+++ examples/example.py
@@ -7,13 +7,14 @@
 import pandas as pd
 from sklearn.linear_model import LinearRegression
 from sklearn.preprocessing import StandardScaler
+from sklearn.feature_selection import SelectKBest, f_regression
 
 # From pyWATTS the pipeline is imported
 from pywatts.callbacks import LinePlotCallback
 from pywatts.core.computation_mode import ComputationMode
 from pywatts.core.pipeline import Pipeline
 # All modules required for the pipeline are imported
-from pywatts.modules import CalendarExtraction, CalendarFeature, ClockShift, LinearInterpolater, SKLearnWrapper
+from pywatts.modules import CalendarExtraction, CalendarFeature, ClockShift, LinearInterpolater, SKLearnWrapper, Sampler
 from pywatts.summaries import RMSE
 
 # The main function is where the pipeline is created and run
@@ -21,7 +22,7 @@ if __name__ == "__main__":
     # Create a pipeline
     pipeline = Pipeline(path="../results")
 
-    # Extract dummy calender features, using holidays from Germany
+    # Extract dummy calendar features, using holidays from Germany
     # NOTE: CalendarExtraction can't return multiple features.
     calendar = CalendarExtraction(continent="Europe", country="Germany", features=[CalendarFeature.month,
                                                                                    CalendarFeature.weekday,
@@ -43,28 +44,38 @@ if __name__ == "__main__":
     shift_power_statistics2 = ClockShift(lag=2, name="ClockShift_Lag2"
                                          )(x=scale_power_statistics)
 
+    target_multiple_output = Sampler(24, name="sampled_data")(x=scale_power_statistics)
+
+    # Select features based on F-statistic
+    selected_features = SKLearnWrapper(
+        module=SelectKBest(score_func=f_regression, k=2)
+    )(
+        power_lag1=shift_power_statistics,
+        power_lag2=shift_power_statistics2,
+        calendar=calendar,
+        target=scale_power_statistics,
+    )
+
     # Create a linear regression that uses the lagged values to predict the current value
     # NOTE: SKLearnWrapper has to collect all **kwargs itself and fit it against target.
     #       It is also possible to implement a join/collect class
     regressor_power_statistics = SKLearnWrapper(
         module=LinearRegression(fit_intercept=True)
     )(
-        power_lag1=shift_power_statistics,
-        power_lag2=shift_power_statistics2,
-        calendar=calendar,
-        target=scale_power_statistics,
-        callbacks=[LinePlotCallback('linear_regression')],
+        features=selected_features,
+        target=target_multiple_output,
+        callbacks=[LinePlotCallback("linear_regression")],
     )
 
     # Rescale the predictions to be on the original time scale
     inverse_power_scale = power_scaler(
         x=regressor_power_statistics, computation_mode=ComputationMode.Transform,
-        use_inverse_transform=True, callbacks=[LinePlotCallback('rescale')]
+        use_inverse_transform=True, callbacks=[LinePlotCallback("rescale")]
     )
 
     # Calculate the root mean squared error (RMSE) between the linear regression and the true values
     # save it as csv file
-    rmse = RMSE()(y_hat=inverse_power_scale, y=pipeline["load_power_statistics"])
+    rmse = RMSE()(y_hat=inverse_power_scale, y=target_multiple_output)
 
     # Now, the pipeline is complete so we can run it and explore the results
     # Start the pipeline